requirements.txt
Flask==1.1.2
gunicorn==20.0.4
numpy==1.19.5
opencv-python-headless
opencv-contrib-python-headless
gluoncv
mxnet
cv2-tools